requirements.txt
Flask==3.0.2
Flask_JWT_Extended==4.6.0
flask_restx==1.3.0
flask_sqlalchemy==3.1.1
Flask_Testing==0.8.1
pandas==2.2.1
PyJWT==2.8.0
PyJWT==2.8.0
SQLAlchemy==2.0.23
vaderSentiment==3.3.2
vaderSentiment==3.3.2
Werkzeug==3.0.1
